+++ scalajvm/app/models/data/Quote.scala
@@ -1,7 +1,7 @@
-package models
+package models.data
 
-import scalikejdbc._
 import org.joda.time.DateTime
+import scalikejdbc._
 
 case class Quote(id: Long, time: DateTime, content: Option[String], rating: Int)
 object Quote extends SQLSyntaxSupport[Quote] {
