--- scalajvm/app/models/forms/SuggestQuoteData.scala
@@ -1,5 +1,0 @@
-package models.forms
-
-case class SuggestQuoteData(content: String, reCaptchaChallenge: String, reCapthaResponse: String) {
-  def toMap = Map("content" -> content)
-}
