+++ scalajvm/app/models/forms/SuggestQuoteForm.scala
@@ -1,0 +1,5 @@
+package models.forms
+
+case class SuggestQuoteForm(content: String, reCaptchaChallenge: String, reCapthaResponse: String) {
+  def toMap = Map("content" -> content)
+}
