+++ scalajvm/app/models/queries/QuoteQueries.scala
@@ -1,5 +1,6 @@
-package models
+package models.queries
 
+import models.data.Quote
 import org.joda.time.DateTime
 import scalikejdbc._
 
@@ -48,15 +49,5 @@ object QuoteQueries {
     }.update().apply() != 0
   }
 
-  def updateRating(increment: Int)(id: Long): Int = {
-    val q = Quote.column
-    DB localTx { implicit session =>
-      withSQL {
-        update(Quote).set(
-          q.rating -> sqls"${q.rating} + $increment"
-        ).where.eq(q.id, id)
-          .append(sqls"returning ${q.rating}")
-      }.map(rs => rs.int(1)).first().apply().getOrElse(0)
-    }
-  }
+
 }
