+++ scalajvm/app/views/main.scala.html
@@ -16,7 +16,7 @@
             <a href="/"><span class="highlight">L</span>og<span class="highlight">L</span>ist</a>
         </div>
 	    <ul class="plate menu center-content">
-		    <li><a href="@routes.UserController.newQuote()">Добавить цитату</a></li>
+		    <li><a href="@routes.SuggestedQuotes.newQuote()">Добавить цитату</a></li>
 	    </ul>
         <div class="content center-content">@content</div>
         <div class="footer center-content"><span class="highlight"><a href="https://github.com/codingteam/loglist">codingteam</a> ©</span> 2014</div>
