+++ scalajvm/app/views/quote.scala.html
@@ -1,4 +1,4 @@
-@(quote: models.Quote)
+@(quote: models.data.Quote)
 
 @main("Log List") {
   <div class="center-content quotes">
