+++ scalajvm/app/views/voting/ratingPanel.scala.html
@@ -1,0 +1,7 @@
+@(quote: models.data.Quote)
+
+<span class="quote-rating" data-id="@quote.id">
+    <span class="quote-button plus">+</span>
+	<span class="quote-rating-value">@quote.rating</span>
+    <span class="quote-button minus">–</span>
+</span>
