+++ scalajvm/conf/routes
@@ -3,14 +3,14 @@
 # ~~~~
 
 # Main routes
-GET           /                                  controllers.UserController.index
-GET           /quote/new                         controllers.UserController.newQuote
-POST          /quote/new                         controllers.UserController.addQuote
-GET           /quote/$id<[0-9]+>                 controllers.UserController.quote(id: String)
-POST          /quote/$id<[0-9]+>/upvote          controllers.UserController.vote(id: String, up: Boolean = true)
-POST          /quote/$id<[0-9]+>/downvote        controllers.UserController.vote(id: String, up: Boolean = false)
-POST          /quote                             controllers.AdminController.addQuote
-DELETE        /quote/$id<[0-9]+>                 controllers.AdminController.deleteQuote(id: String)
+GET           /                                  controllers.Quotes.index
+GET           /quote/new                         controllers.SuggestedQuotes.newQuote
+POST          /quote/new                         controllers.SuggestedQuotes.addQuote
+GET           /quote/$id<[0-9]+>                 controllers.Quotes.quote(id: String)
+POST          /quote/$id<[0-9]+>/upvote          controllers.Voting.vote(id: String, up: Boolean = true)
+POST          /quote/$id<[0-9]+>/downvote        controllers.Voting.vote(id: String, up: Boolean = false)
+POST          /quote                             controllers.Quotes.addQuote
+DELETE        /quote/$id<[0-9]+>                 controllers.Quotes.deleteQuote(id: String)
 
 
 # Map static resources from the /public folder to the /assets URL path
